% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soe_data.R
\name{get_soe_data}
\alias{get_soe_data}
\title{Function to get from ALA data in a standard format}
\usage{
get_soe_data(type, taxa, spatial, temporal)
}
\description{
Primarily internal function for getting data in correct format
}
