% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibra_map.R
\name{ibra_map}
\alias{ibra_map}
\title{Map of IBRA regions}
\format{
sf spatial object
source \url{http://www.environment.gov.au/metadataexplorer/explorer.jsp}
}
\description{
IBRA map
}
\details{
Simplified map of IBRA regions
}
